pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "nova-agent"
version = "0.1.0"
description = "Self-evolving AI agent with compounding knowledge, two-tier memory, cron scheduling, and autonomous self-improvement"
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"
authors = [{name = "Nova Agent Contributors"}]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "anthropic>=0.40.0",
    "openai>=1.30.0",
    "croniter>=2.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-cov>=4.0"]
telegram = ["python-telegram-bot>=20.0"]
discord = ["discord.py>=2.3"]
wechat = ["itchat>=1.3", "pycryptodome", "qrcode", "requests"]
feishu = ["lark-oapi>=1.4"]
qq = ["qq-botpy>=1.3"]
dingtalk = ["dingtalk-stream>=1.0"]
web = ["flask>=3.0"]
gateway = [
    "python-telegram-bot>=20.0",
    "discord.py>=2.3",
    "itchat>=1.3",
    "pycryptodome",
    "qrcode",
    "requests",
    "lark-oapi>=1.4",
    "qq-botpy>=1.3",
    "dingtalk-stream>=1.0",
    "flask>=3.0",
]
all = [
    "python-telegram-bot>=20.0",
    "discord.py>=2.3",
    "itchat>=1.3",
    "pycryptodome",
    "qrcode",
    "requests",
    "lark-oapi>=1.4",
    "qq-botpy>=1.3",
    "dingtalk-stream>=1.0",
    "flask>=3.0",
]

[project.scripts]
nova = "nova.main:main"
nova-tg = "nova.gateway.telegram:main"
nova-discord = "nova.gateway.discord:main"
nova-wechat = "nova.gateway.wechat:main"
nova-feishu = "nova.gateway.feishu:main"
nova-qq = "nova.gateway.qq:main"
nova-dingtalk = "nova.gateway.dingtalk:main"
nova-web = "nova.gateway.webhook:main_minimal"
nova-schedule = "nova.scheduler:cli"

[tool.setuptools.packages.find]
where = ["."]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]